requirements.txt
flask
python-decouple
flask_cors